ref: move variable for testing to its own file, so it can be reused
